AWS/difuminado_caras/requirements.txt
boto3
Flask
python-dotenv
opencv-python
numpy